Refactor fetching target clusters to its own common reducer
