+++ app/javascript/common/constants.js
@@ -34,11 +34,3 @@ export const V2V_TARGET_PROVIDERS = [
     type: 'ManageIQ::Providers::Openstack::CloudManager'
   }
 ];
-
-export const FETCH_TARGET_COMPUTE_URLS = {
-  [RHV]:
-    '/api/clusters?expand=resources' +
-    '&attributes=ext_management_system.emstype,v_parent_datacenter,ext_management_system.name,hosts' +
-    '&filter[]=ext_management_system.emstype=rhevm',
-  [OPENSTACK]: '/api/cloud_tenants?expand=resources&attributes=ext_management_system.name,ext_management_system.id'
-};
