+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardClustersStep/MappingWizardClustersStepActions.js
@@ -1,10 +1,6 @@
 import URI from 'urijs';
 import API from '../../../../../../../../common/API';
-import {
-  FETCH_V2V_SOURCE_CLUSTERS,
-  FETCH_V2V_TARGET_CLUSTERS,
-  QUERY_V2V_PROVIDERS
-} from './MappingWizardClustersStepConstants';
+import { FETCH_V2V_SOURCE_CLUSTERS, QUERY_V2V_PROVIDERS } from './MappingWizardClustersStepConstants';
 
 export { showAlertAction, hideAlertAction } from '../../MappingWizardActions';
 
@@ -17,17 +13,6 @@ const _getSourceClustersActionCreator = url => dispatch =>
 export const fetchSourceClustersAction = url => {
   const uri = new URI(url);
   return _getSourceClustersActionCreator(uri.toString());
-};
-
-const _getTargetClustersActionCreator = url => dispatch =>
-  dispatch({
-    type: FETCH_V2V_TARGET_CLUSTERS,
-    payload: API.get(url)
-  });
-
-export const fetchTargetClustersAction = url => {
-  const uri = new URI(url);
-  return _getTargetClustersActionCreator(uri.toString());
 };
 
 // ****************************************************************************
