+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardClustersStep/__tests__/MappingWizardClusterStepActions.test.js
@@ -1,10 +1,6 @@
 import { mockStore } from '../../../../../../common/testReduxHelpers';
 import * as actions from '../MappingWizardClustersStepActions';
-import {
-  initialState,
-  requestSourceClustersData,
-  requestTargetClustersData
-} from '../mappingWizardClustersStep.fixtures';
+import { initialState, requestSourceClustersData } from '../mappingWizardClustersStep.fixtures';
 import { mockRequest, mockReset } from '../../../../../../../../../common/mockRequests';
 
 const store = mockStore(initialState);
@@ -32,27 +28,6 @@ describe('mappingWizard actions', () => {
       status: 404
     });
     return store.dispatch(actions.fetchSourceClustersAction(fetchSourceClustersUrl)).catch(() => {
-      expect(store.getActions()).toMatchSnapshot();
-    });
-  });
-
-  it('should fetch target clusters and return PENDING and FULFILLED action', () => {
-    const { fetchTargetClustersUrl } = requestTargetClustersData;
-    mockRequest({
-      url: fetchTargetClustersUrl,
-      status: 200
-    });
-    return store.dispatch(actions.fetchTargetClustersAction(fetchTargetClustersUrl)).then(() => {
-      expect(store.getActions()).toMatchSnapshot();
-    });
-  });
-  it('should fetch target clusters and return PENDING and REJECTED action', () => {
-    const { fetchTargetClustersUrl } = requestTargetClustersData;
-    mockRequest({
-      url: fetchTargetClustersUrl,
-      status: 404
-    });
-    return store.dispatch(actions.fetchTargetClustersAction(fetchTargetClustersUrl)).catch(() => {
       expect(store.getActions()).toMatchSnapshot();
     });
   });
