+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardClustersStep/__tests__/MappingWizardClustersStepReducers.test.js
@@ -1,11 +1,7 @@
 import Immutable from 'seamless-immutable';
 import clustersReducer, { initialState } from '../MappingWizardClustersStepReducer';
-import {
-  FETCH_V2V_SOURCE_CLUSTERS,
-  FETCH_V2V_TARGET_CLUSTERS,
-  QUERY_V2V_PROVIDERS
-} from '../MappingWizardClustersStepConstants';
-import { sourceClusters, targetClusters } from '../mappingWizardClustersStep.fixtures';
+import { FETCH_V2V_SOURCE_CLUSTERS, QUERY_V2V_PROVIDERS } from '../MappingWizardClustersStepConstants';
+import { sourceClusters } from '../mappingWizardClustersStep.fixtures';
 
 test('sets default state', () => {
   const action = { type: '@@INIT' };
@@ -72,72 +68,6 @@ describe('fetching source clusters', () => {
       };
       const action = {
         type: `${FETCH_V2V_SOURCE_CLUSTERS}_FULFILLED`,
-        payload
-      };
-      const state = clustersReducer(prevState, action);
-
-      expect(state).toMatchSnapshot();
-    });
-  });
-});
-
-describe('fetching target clusters', () => {
-  test('is pending', () => {
-    const action = {
-      type: `${FETCH_V2V_TARGET_CLUSTERS}_PENDING`
-    };
-    const prevState = initialState.set('isRejectedTargetClusters', true);
-    const state = clustersReducer(prevState, action);
-
-    expect(state).toMatchSnapshot();
-  });
-
-  test('is rejected', () => {
-    const action = {
-      type: `${FETCH_V2V_TARGET_CLUSTERS}_REJECTED`,
-      payload: 'error'
-    };
-    const prevState = initialState.set('isFetchingTargetClusters', true);
-    const state = clustersReducer(prevState, action);
-
-    expect(state).toMatchSnapshot();
-  });
-
-  describe('is successful,', () => {
-    let prevState;
-    beforeEach(() => {
-      prevState = initialState.set('isRejectedTargetClusters', true).set('isFetchingTargetClusters', true);
-    });
-    test('and there is data', () => {
-      const { resources } = targetClusters;
-      const payload = {
-        data: {
-          ...targetClusters,
-          resources: [
-            {
-              ...resources
-            }
-          ]
-        }
-      };
-      const action = {
-        type: `${FETCH_V2V_TARGET_CLUSTERS}_FULFILLED`,
-        payload
-      };
-      const state = clustersReducer(prevState, action);
-
-      expect(state).toMatchSnapshot();
-    });
-
-    test('and there is no data', () => {
-      const payload = {
-        data: {
-          ...targetClusters,
-          resources: []
-        }
-      };
-      const action = {
-        type: `${FETCH_V2V_TARGET_CLUSTERS}_FULFILLED`,
         payload
       };
       const state = clustersReducer(prevState, action);
