+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardClustersStep/components/ClustersStepForm/clustersStepForm.fixtures.js
@@ -1,4 +1,5 @@
-import { sourceClusters, targetClusters } from '../../mappingWizardClustersStep.fixtures';
+import { sourceClusters } from '../../mappingWizardClustersStep.fixtures';
+import { targetClusters } from '../../../../../../../../../../redux/common/targetResources/targetResources.fixtures';
 
 export const srcClusters = sourceClusters.resources;
 export const tgtClusters = targetClusters.resources;
