+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardClustersStep/index.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux';
 
 import MappingWizardClustersStep from './MappingWizardClustersStep';
 import * as MappingWizardClustersStepActions from './MappingWizardClustersStepActions';
+import * as TargetResourcesActions from '../../../../../../../../redux/common/targetResources/targetResourcesActions';
 import { createClusterMappings } from './components/ClustersStepForm/helpers';
 import { getTransformationMappingItemsBySourceType } from '../../helpers';
 import { TRANSFORMATION_MAPPING_ITEM_SOURCE_TYPES, OPENSTACK, RHV } from '../../MappingWizardConstants';
@@ -12,10 +13,11 @@ import reducer from './MappingWizardClustersStepReducer';
 export const reducers = { mappingWizardClustersStep: reducer };
 
 const mapStateToProps = (
-  { mappingWizardClustersStep, mappingWizardGeneralStep: { editingMapping, conversionHosts }, form },
+  { mappingWizardClustersStep, mappingWizardGeneralStep: { editingMapping, conversionHosts }, form, targetResources },
   ownProps
 ) => ({
   ...mappingWizardClustersStep,
+  ...targetResources,
   ...ownProps.data,
   targetProvider: form.mappingWizardGeneralStep.values.targetProvider,
   rhvConversionHosts: conversionHostsFilter(conversionHosts, RHV),
@@ -24,7 +26,7 @@ const mapStateToProps = (
     clusterMappings: editingMapping
       ? createClusterMappings(
           getTransformationMappingItemsBySourceType(TRANSFORMATION_MAPPING_ITEM_SOURCE_TYPES.cluster, editingMapping),
-          mappingWizardClustersStep.targetClusters,
+          targetResources.targetClusters,
           mappingWizardClustersStep.sourceClusters
         )
       : []
@@ -37,6 +39,6 @@ const mergeProps = (stateProps, dispatchProps, ownProps) => Object.assign(stateP
 
 export default connect(
   mapStateToProps,
-  MappingWizardClustersStepActions,
+  { ...MappingWizardClustersStepActions, ...TargetResourcesActions },
   mergeProps
 )(MappingWizardClustersStep);
