+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardDatastoresStep/__tests__/MappingWizardDatastoresStep.test.js
@@ -6,7 +6,8 @@ import { reducer as formReducer } from 'redux-form';
 import { generateStore } from '../../../../../../common/testReduxHelpers';
 import { reducers } from '../index';
 import MappingWizardDatastoresStep from '../MappingWizardDatastoresStep';
-import { sourceClusters, cloudTenants } from '../../MappingWizardClustersStep/mappingWizardClustersStep.fixtures';
+import { sourceClusters } from '../../MappingWizardClustersStep/mappingWizardClustersStep.fixtures';
+import { cloudTenants } from '../../../../../../../../../redux/common/targetResources/targetResources.fixtures';
 import { constructClusterMappings } from '../../../mappingWizardTestHelpers';
 import { FETCH_STORAGE_URLS } from '../MappingWizardDatastoresStepConstants';
 import { V2V_TARGET_PROVIDERS } from '../../../MappingWizardConstants';
