+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardNetworksStep/__tests__/MappingWizardNetworksStep.test.js
@@ -6,7 +6,8 @@ import { reducer as formReducer } from 'redux-form';
 import { generateStore } from '../../../../../../common/testReduxHelpers';
 import { reducers } from '../index';
 import MappingWizardNetworksStep from '../MappingWizardNetworksStep';
-import { sourceClusters, cloudTenants } from '../../MappingWizardClustersStep/mappingWizardClustersStep.fixtures';
+import { sourceClusters } from '../../MappingWizardClustersStep/mappingWizardClustersStep.fixtures';
+import { cloudTenants } from '../../../../../../../../../redux/common/targetResources/targetResources.fixtures';
 import { constructClusterMappings } from '../../../mappingWizardTestHelpers';
 import { V2V_TARGET_PROVIDERS } from '../../../MappingWizardConstants';
 import { FETCH_NETWORK_URLS } from '../MappingWizardNetworksStepConstants';
