+++ app/javascript/react/screens/App/Settings/SettingsActions.js
@@ -9,8 +9,7 @@ import {
   FETCH_V2V_CONVERSION_HOSTS,
   SHOW_V2V_CONVERSION_HOST_WIZARD,
   HIDE_V2V_CONVERSION_HOST_WIZARD,
-  V2V_CONVERSION_HOST_WIZARD_EXITED,
-  FETCH_V2V_TARGET_COMPUTE_RESOURCES
+  V2V_CONVERSION_HOST_WIZARD_EXITED
 } from './SettingsConstants';
 import { getApiSettingsFromFormValues } from './helpers';
 import { stepIDs } from './screens/ConversionHostsSettings/components/ConversionHostWizard/ConversionHostWizardConstants';
@@ -75,15 +74,4 @@ export const conversionHostWizardExitedAction = () => dispatch => {
   Object.values(stepIDs).forEach(formName => {
     dispatch(reset(formName));
   });
-};
-
-const _getTargetComputeResourcesActionCreator = url => dispatch =>
-  dispatch({
-    type: FETCH_V2V_TARGET_COMPUTE_RESOURCES,
-    payload: API.get(url)
-  });
-
-export const fetchTargetComputeResourcesAction = url => {
-  const uri = new URI(url);
-  return _getTargetComputeResourcesActionCreator(uri.toString());
 };
