+++ app/javascript/react/screens/App/Settings/screens/ConversionHostsSettings/components/ConversionHostWizard/ConversionHostWizardLocationStep/index.js
@@ -1,25 +1,25 @@
 import { connect } from 'react-redux';
 import ConversionHostWizardLocationStep from './ConversionHostWizardLocationStep';
 
-import * as SettingsActions from '../../../../../SettingsActions';
+import * as TargetResourceActions from '../../../../../../../../../redux/common/targetResources/targetResourcesActions';
 
 import { stepIDs } from '../ConversionHostWizardConstants';
 
 const mapStateToProps = ({
   form,
   providers: { providers },
-  settings: { isFetchingTargetComputeResources, targetComputeResources }
+  targetResources: { isFetchingTargetClusters, targetClusters }
 }) => ({
   locationStepForm: form[stepIDs.locationStep],
   providers,
-  isFetchingTargetComputeResources,
-  targetComputeResources
+  isFetchingTargetClusters,
+  targetClusters
 });
 
 const mergeProps = (stateProps, dispatchProps, ownProps) => Object.assign(stateProps, ownProps.data, dispatchProps);
 
 export default connect(
   mapStateToProps,
-  SettingsActions,
+  TargetResourceActions,
   mergeProps
 )(ConversionHostWizardLocationStep);
